backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
dj-database-url==2.1.0
Django==4.2.21
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-extensions==4.1
django-js-asset==2.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
numpy==2.0.2
oauth2client==4.1.3
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg[binary]==3.2.9
pycryptodome==3.23.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyDrive==1.3.1
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.6.0
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
Brotli>=1.0,<2